+++ Game/Objs/Obj_Structure_ParticleAccelerator.cs
@@ -157,7 +157,7 @@ namespace Somnium.Game {
 		// Function from file: particle_accelerator.dm
 		public override bool ex_act( double? severity = null, dynamic child = null ) {
 			
-			switch ((double?)( severity )) {
+			switch ((int?)( severity )) {
 				case 1:
 					GlobalFuncs.qdel( this );
 					return false;
